+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This file is used to list changes made in each version of the resolver cookbook.
 ## Unreleased
 
 - Sous Chefs Adoption
+- Add inspec Tests
+- Remove Deprecated Platforms
+- Change ci to point to Sous Chefs
 
 ## 2.1.0 (2018-07-18)